Remove Debugger.NodeState.Web and dummy assigns
